# workforce_api_openapi.json
{"openapi":"3.1.0","info":{"title":"NuiFlo WorkForce API","description":"\n    # NuiFlo WorkForce API\n    \n    AI Team Management Platform powered by CrewAI. Build, deploy, and scale virtual teams of AI agents for custom workflows.\n    \n    ## Features\n    \n    * **Team Management**: Create and manage AI teams with custom roles\n    * **Role Configuration**: Define agent expertise levels and LLM models  \n    * **Budget Tracking**: Monitor spending and set monthly budgets\n    * **Real-time Execution**: Execute teams using CrewAI framework\n    * **Status Monitoring**: Track team performance and utilization\n    \n    ## Authentication\n    \n    Currently using dummy authentication (user_id=1). Production authentication will be implemented.\n    \n    ## Usage for Frontend Integration\n    \n    1. **Create Team**: `POST /api/v1/teams/` with team details and roles\n    2. **List Teams**: `GET /api/v1/teams/` to get all user teams\n    3. **Get Status**: `GET /api/v1/teams/{id}/status` for detailed metrics\n    4. **Execute Team**: `POST /api/v1/teams/{id}/execute` to run AI workflows\n    \n    ## Contact\n    \n    * **Developer**: NuiFlo Team\n    * **Email**: team@nuiflo.com\n    * **Repository**: [GitHub](https://github.com/nuiflo/workforce)\n    ","contact":{"name":"NuiFlo Team","email":"team@nuiflo.com"},"license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"version":"0.1.0"},"servers":[{"url":"http://localhost:8000","description":"Development server"},{"url":"https://api.nuiflo.com","description":"Production server (future)"}],"paths":{"/health/ping":{"get":{"tags":["health"],"summary":"Ping","description":"Basic ping endpoint.","operationId":"ping_health_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/status":{"get":{"tags":["health"],"summary":"Health Status","description":"Comprehensive health check.","operationId":"health_status_health_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/teams/":{"post":{"tags":["teams"],"summary":"Create Team","description":"Create a new team with roles.\n\nArgs:\n    team_data: Team creation data\n    db: Database session\n    \nReturns:\n    Created team information","operationId":"create_team_api_v1_teams__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["teams"],"summary":"List Teams","description":"List teams.\n\nArgs:\n    user_id: Optional user ID filter\n    db: Database session\n    \nReturns:\n    List of teams","operationId":"list_teams_api_v1_teams__get","parameters":[{"name":"user_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Filter by user ID","title":"User Id"},"description":"Filter by user ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TeamResponse"},"title":"Response List Teams Api V1 Teams  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/teams/{team_id}":{"get":{"tags":["teams"],"summary":"Get Team","description":"Get team by ID.\n\nArgs:\n    team_id: Team ID\n    db: Database session\n    \nReturns:\n    Team information","operationId":"get_team_api_v1_teams__team_id__get","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"integer","title":"Team Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["teams"],"summary":"Update Team","description":"Update team information.\n\nArgs:\n    team_id: Team ID\n    team_data: Team update data\n    db: Database session\n    \nReturns:\n    Updated team information","operationId":"update_team_api_v1_teams__team_id__put","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"integer","title":"Team Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["teams"],"summary":"Delete Team","description":"Delete a team.\n\nArgs:\n    team_id: Team ID\n    db: Database session","operationId":"delete_team_api_v1_teams__team_id__delete","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"integer","title":"Team Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/teams/{team_id}/execute":{"post":{"tags":["teams"],"summary":"Execute Team","description":"Execute a team using CrewAI.\n\nArgs:\n    team_id: Team ID to execute\n    execution_data: Optional execution parameters\n    db: Database session\n    \nReturns:\n    Execution results and metrics","operationId":"execute_team_api_v1_teams__team_id__execute_post","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"integer","title":"Team Id"}}],"requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/TeamExecute"},{"type":"null"}],"title":"Execution Data"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecutionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/teams/{team_id}/status":{"get":{"tags":["teams"],"summary":"Get Team Status","description":"Get detailed team status information.\n\nArgs:\n    team_id: Team ID\n    db: Database session\n    \nReturns:\n    Team status information","operationId":"get_team_status_api_v1_teams__team_id__status_get","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"integer","title":"Team Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Team Status Api V1 Teams  Team Id  Status Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["root"],"summary":"Root","description":"Welcome endpoint with API information.\n\nReturns basic information about the NuiFlo WorkForce API.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ExecutionResponse":{"properties":{"result":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result"},"metrics":{"additionalProperties":true,"type":"object","title":"Metrics"},"success":{"type":"boolean","title":"Success"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"},"team_execution_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Team Execution Id"}},"type":"object","required":["result","metrics","success","error","team_execution_id"],"title":"ExecutionResponse"},"ExpertiseLevel":{"type":"string","enum":["junior","intermediate","senior","expert"],"title":"ExpertiseLevel"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RoleCreate":{"properties":{"title":{"type":"string","maxLength":100,"minLength":1,"title":"Title"},"description":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Description"},"expertise":{"$ref":"#/components/schemas/ExpertiseLevel"},"llm_model":{"type":"string","maxLength":50,"title":"Llm Model","default":"gpt-3.5-turbo"},"llm_config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Llm Config"},"agent_config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Agent Config"},"is_active":{"type":"boolean","title":"Is Active","default":true}},"type":"object","required":["title","expertise"],"title":"RoleCreate"},"RoleResponse":{"properties":{"id":{"type":"integer","title":"Id"},"team_id":{"type":"integer","title":"Team Id"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"expertise":{"$ref":"#/components/schemas/ExpertiseLevel"},"llm_model":{"type":"string","title":"Llm Model"},"llm_config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Llm Config"},"agent_config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Agent Config"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"}},"type":"object","required":["id","team_id","title","description","expertise","llm_model","llm_config","agent_config","is_active","created_at","updated_at"],"title":"RoleResponse"},"TeamCreate":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name"},"description":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Description"},"monthly_budget":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string"}],"title":"Monthly Budget","description":"Monthly budget in USD"},"roles":{"items":{"$ref":"#/components/schemas/RoleCreate"},"type":"array","maxItems":10,"title":"Roles"}},"type":"object","required":["name","monthly_budget"],"title":"TeamCreate"},"TeamExecute":{"properties":{"inputs":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Inputs","description":"Execution inputs"}},"type":"object","title":"TeamExecute"},"TeamResponse":{"properties":{"id":{"type":"integer","title":"Id"},"owner_id":{"type":"integer","title":"Owner Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"monthly_budget":{"type":"string","title":"Monthly Budget"},"current_spend":{"type":"string","title":"Current Spend"},"status":{"type":"string","title":"Status"},"last_executed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Executed At"},"created_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created At"},"updated_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Updated At"},"roles":{"items":{"$ref":"#/components/schemas/RoleResponse"},"type":"array","title":"Roles","default":[]}},"type":"object","required":["id","owner_id","name","description","monthly_budget","current_spend","status","last_executed_at","created_at","updated_at"],"title":"TeamResponse"},"TeamUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Description"},"monthly_budget":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string"},{"type":"null"}],"title":"Monthly Budget"}},"type":"object","title":"TeamUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"health","description":"Health check endpoints for monitoring service status"},{"name":"teams","description":"Team management operations - the core of the AI workforce platform"}]}